feat(event_tracker): cleanup PE-4982
