+++ lib/pages/drive_detail/drive_detail_page.dart
@@ -34,7 +34,7 @@ import 'package:ardrive/utils/filesize.dart';
 import 'package:ardrive/utils/logger/logger.dart';
 import 'package:ardrive/utils/mobile_screen_orientation.dart';
 import 'package:ardrive/utils/mobile_status_bar.dart';
-import 'package:ardrive/utils/plausible_event_tracker.dart';
+import 'package:ardrive/utils/plausible_event_tracker/plausible_event_tracker.dart';
 import 'package:ardrive/utils/size_constants.dart';
 import 'package:ardrive/utils/user_utils.dart';
 import 'package:ardrive_io/ardrive_io.dart';
@@ -80,10 +80,10 @@ class _DriveDetailPageState extends State<DriveDetailPage> {
 
     if (widget.anonymouslyShowDriveDetail) {
       PlausibleEventTracker.trackPageview(
-          event: PlausiblePageView.fileExplorerNonLoggedInUser);
+          page: PlausiblePageView.fileExplorerNonLoggedInUser);
     } else {
       PlausibleEventTracker.trackPageview(
-          event: PlausiblePageView.fileExplorerLoggedInUser);
+          page: PlausiblePageView.fileExplorerLoggedInUser);
     }
   }
