+++ lib/pages/no_drives/no_drives_page.dart
@@ -3,7 +3,7 @@ import 'package:ardrive/blocs/drive_detail/drive_detail_cubit.dart';
 import 'package:ardrive/components/profile_card.dart';
 import 'package:ardrive/pages/drive_detail/drive_detail_page.dart';
 import 'package:ardrive/utils/app_localizations_wrapper.dart';
-import 'package:ardrive/utils/plausible_event_tracker.dart';
+import 'package:ardrive/utils/plausible_event_tracker/plausible_event_tracker.dart';
 import 'package:ardrive_ui/ardrive_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -20,11 +20,11 @@ class NoDrivesPage extends StatelessWidget {
   }) : super(key: key) {
     if (anonymouslyShowDriveDetail) {
       PlausibleEventTracker.trackPageview(
-        event: PlausiblePageView.fileExplorerNonLoggedInUser,
+        page: PlausiblePageView.fileExplorerNonLoggedInUser,
       );
     } else {
       PlausibleEventTracker.trackPageview(
-        event: PlausiblePageView.fileExplorerNewUserEmpty,
+        page: PlausiblePageView.fileExplorerNewUserEmpty,
       );
     }
   }
