+++ lib/turbo/topup/views/topup_modal.dart
@@ -14,7 +14,7 @@ import 'package:ardrive/turbo/topup/views/topup_success_view.dart';
 import 'package:ardrive/turbo/topup/views/turbo_error_view.dart';
 import 'package:ardrive/turbo/turbo.dart';
 import 'package:ardrive/utils/logger/logger.dart';
-import 'package:ardrive/utils/plausible_event_tracker.dart';
+import 'package:ardrive/utils/plausible_event_tracker/plausible_event_tracker.dart';
 import 'package:ardrive/utils/show_general_dialog.dart';
 import 'package:ardrive_ui/ardrive_ui.dart';
 import 'package:flutter/material.dart';
@@ -62,7 +62,7 @@ void showTurboTopupModal(BuildContext context, {Function()? onSuccess}) {
 
   activityTracker.setToppingUp(true);
 
-  PlausibleEventTracker.trackPageview(event: PlausiblePageView.turboTopUpModal);
+  PlausibleEventTracker.trackPageview(page: PlausiblePageView.turboTopUpModal);
 
   showAnimatedDialogWithBuilder(
     context,
@@ -94,13 +94,13 @@ void showTurboTopupModal(BuildContext context, {Function()? onSuccess}) {
     if (turbo.paymentStatus == PaymentStatus.success) {
       logger.d('Turbo payment success');
       PlausibleEventTracker.trackPageview(
-          event: PlausiblePageView.turboTopUpSuccess);
+          page: PlausiblePageView.turboTopUpSuccess);
 
       onSuccess?.call();
     } else {
       logger.d('Turbo payment error');
       PlausibleEventTracker.trackPageview(
-          event: PlausiblePageView.turboTopUpCancel);
+          page: PlausiblePageView.turboTopUpCancel);
     }
 
     turbo.dispose();
@@ -179,7 +179,7 @@ class _TurboModalState extends State<TurboModal> with TickerProviderStateMixin {
           view = const TopUpEstimationView();
         } else if (state is TurboTopupFlowShowingPaymentFormView) {
           PlausibleEventTracker.trackPageview(
-            event: PlausiblePageView.turboPaymentDetails,
+            page: PlausiblePageView.turboPaymentDetails,
           );
           view = Stack(
             children: [
@@ -203,7 +203,7 @@ class _TurboModalState extends State<TurboModal> with TickerProviderStateMixin {
           );
         } else if (state is TurboTopupFlowShowingPaymentReviewView) {
           PlausibleEventTracker.trackPageview(
-            event: PlausiblePageView.turboPurchaseReview,
+            page: PlausiblePageView.turboPurchaseReview,
           );
           view = Stack(
             children: [
