+++ lib/utils/plausible_event_tracker/plausible_api_data.dart
@@ -1,0 +1,9 @@
+class PlausibleApiData {
+  final Uri api;
+  final String domain;
+
+  PlausibleApiData({
+    required this.api,
+    required this.domain,
+  });
+}
