+++ lib/utils/plausible_event_tracker/plausible_event_data.dart
@@ -1,0 +1,13 @@
+class PlausibleEventData {
+  final String name;
+  final Uri url;
+  String? referrer;
+  Map<String, dynamic>? props;
+
+  PlausibleEventData({
+    required this.name,
+    required this.url,
+    this.referrer,
+    this.props,
+  });
+}
