+++ lib/utils/plausible_event_tracker/plausible_page_view_events.dart
@@ -1,0 +1,30 @@
+enum PlausiblePageView {
+  fileExplorerLoggedInUser,
+  fileExplorerNewUserEmpty,
+  fileExplorerNonLoggedInUser,
+  tutorialsPage1,
+  tutorialsPage2,
+  tutorialsPage3,
+  tutorialSkipped,
+  createAndConfirmPasswordPage,
+  createdAndConfirmedPassword,
+  gettingStartedPage,
+  enterSeedPhrasePage,
+  logout,
+  onboardingPage,
+  sharedFilePage,
+  turboPaymentDetails,
+  turboPurchaseReview,
+  turboTopUpModal,
+  turboTopUpCancel,
+  turboTopUpSuccess,
+  verifySeedPhrasePage,
+  writeDownSeedPhrasePage,
+  walletDownloadPage,
+  walletDownloaded,
+  walletGenerationPage,
+  welcomePage,
+  welcomeBackPage,
+
+  unknown,
+}
